refactor && edit for parser

See merge request en1qc/ifj23!25

+++ tests/unit/symtable_test05/main.c
@@ -38,7 +38,7 @@ int main()
 
     assert(err == false);
 
-    uint8_t error = 255;
+    unsigned int error = 255;
     //insert new param
     assert(add_param(&global_sym_table, &item, &param, &err) == 0);
     //set label for param
